MAJOR cleanup affecting essentially every class in this project. The biggest change is likely the addition of 'final' keywords to most if not all method parameters that can legitimately have the keyword (obviously, parameters which are changed within the method body cannot and are not marked 'final' although changing parameters within the method body is something that I really try to avoid doing).

+++ src/org/apache/commons/text/similarity/LevenshteinDistance.java
@@ -71,7 +71,7 @@ public class LevenshteinDistance implements EditDistance<Integer> {
      *        This may not be negative.
      */
     public LevenshteinDistance(final Integer threshold) {
-        if (threshold != null && threshold < 0) {
+        if (threshold != null && threshold.intValue() < 0) {
             throw new IllegalArgumentException("Threshold must not be negative");
         }
         this.threshold = threshold;
@@ -112,7 +112,7 @@ public class LevenshteinDistance implements EditDistance<Integer> {
     @Override
     public Integer apply(final CharSequence left, final CharSequence right) {
         if (threshold != null) {
-            return limitedCompare(left, right, threshold);
+            return limitedCompare( left, right, threshold.intValue() );
         }
         return unlimitedCompare(left, right);
     }
